minor tweaks to fix compilation warnings with -Wall

+++ common/binding.h
@@ -1,11 +1,11 @@
-/* 
+/*
  * File:   binding.h
  * Author: tabrown
  *
  * Created on June 23, 2016, 6:52 PM
- * 
+ *
  * Used to configure and implement a thread binding/pinning policy.
- * 
+ *
  * Instructions:
  * 1. invoke binding_configurePolicy, passing the number of logical processors.
  * 2. invoke binding_parseCustom, passing a string that describes the desired
@@ -54,50 +54,50 @@ static unsigned parseToken(std::string argv, int ix) {
     //      INT-INT
     // and is either followed by "," is the end of the string.
     // we first determine which is the case
-    
+
     // read first INT
     int ix2 = ix;
-    while (ix2 < argv.size() && argv[ix2] != '.') ++ix2;
+    while (ix2 < (int) argv.size() && argv[ix2] != '.') ++ix2;
     std::string token = argv.substr(ix, ix2-ix+1);
     int a = atoi(token.c_str());
-    
+
     // check if the token is of the first form: INT
     ix = ix+digits(a);              // first character AFTER first INT
-    if (ix >= argv.size() || argv[ix] == '.') {
-        
+    if (ix >= (int) argv.size() || argv[ix] == '.') {
+
         // add single binding
         //cout<<"a="<<a<<std::endl;
         customBinding[numCustomBindings++] = a;
         assert(numCustomBindings <= MAX_THREADS_POW2);
-    
+
     // token is of the second form: INT-INT
     } else {
         assert(argv[ix] == '-');
         ++ix;                       // skip '-'
-        
+
         // read second INT
         token = argv.substr(ix, ix2-ix+1);
         int b = atoi(token.c_str());
         //cout<<"a="<<a<<" b="<<b<<std::endl;
-        
+
         // add range of bindings
         for (int i=a;i<=b;++i) {
             customBinding[numCustomBindings++] = i;
             assert(numCustomBindings <= MAX_THREADS_POW2);
         }
-        
+
         ix = ix+digits(b);          // first character AFTER second INT
     }
     // note: ix is the first character AFTER the last INT in the token
     // this is either a comma ('.') or the end of the std::string argv.
-    return (ix >= argv.size() ? argv.size() : ix+1 /* skip '.' */);
+    return (ix >= (int) argv.size() ? argv.size() : ix+1 /* skip '.' */);
 }
 
 // argv contains a custom thread binding pattern, e.g., "1,2,3,8-11,4-7,0"
 // threads will be bound according to this binding
 void binding_parseCustom(std::string argv) {
     numCustomBindings = 0;
-    
+
     unsigned ix = 0;
     while (ix < argv.size()) {
         ix = parseToken(argv, ix);
